refactor(web): ハンドラファクトリーを server 直下のフラットファイルへ移動
server/handler/ は横断的なファクトリー1ファイルのみを抱えており、
feature 配下の server/<feature>/handler/（個別エンドポイント実装）と
名前が衝突して意味が紛らわしかった。lone なディレクトリを解消し、
server/handler-factory.ts に移す。
